STV_ESCALA_v1: escalas fraccionales de render (x1.25/x1.5/x1.75)
God of War esta limitado por fill de GPU a x2 (960x544). Las escalas
intermedias recortan el fill para llegar a 60 fps sin caer a x1.

Palanca: prop debug.stv.escala / env STV_ESCALA con valores "125"/"150"/
"175"; cualquier otro valor o ausencia = OFF (upstream exacto). Se lee UNA
vez por proceso en UpdateRenderSize (static const): conmutar exige
reiniciar el juego. Solo pisa la config x2 del usuario; con x1/x3 se
ignora y se avisa por log ("STV: escala IGNORADA"). Activa:
"STV: escala ACTIVA 1.5 (720x408) (STV_ESCALA_v1)".

El cambio de fondo: FramebufferManagerCommon::renderScaleFactor_ (int) y
VirtualFramebuffer::renderScaleFactor (u8) pasan a float. Los factores
1.25/1.5/1.75 son exactos en binario, asi que las comparaciones == 1.0f
y los redondeos quedan deterministas. Reglas aplicadas en los ~50 usos:

- Tamanos de framebuffer y rects de blits/copias: StvEscalarDim
  (floorf(v*f + 0.5f)), el redondeo UNICO, identico en src y dst de cada
  blit — sin drift entre iteraciones del feedback (self-texturing) de
  GoW. floorf y no cast: el cast trunca hacia cero y con offsets
  negativos (CopySource) romperia la identidad de productos exactos.
- Comparaciones == 1 / != 1 -> 1.0f (exacto).
- La cadena del scaleFactor de los blits 2D (BlitUsingRaster ->
  Draw2D::Blit -> DrawStrip2D) pasa de int a float: siempre termino en
  un uniform float del shader de deswizzle, la fraccion no se pierde
  (con x1.5 la franja del deswizzle mide 6.0 texeles exactos).
- Readbacks (GoW no los usa): factor 1.0f literal y rects con el helper.
- TextureCacheCommon (self-texturing/depal, DELICADOS): rects a texel
  con el helper; los UV bounds del depal quedan float como en upstream.
- Debug (GPUDebugBuffer, tempfbo name, ImGe): casts explicitos,
  cosmeticos.
- PSP_CoreParameter().renderScaleFactor (config entera global) intacto.
- VirtualFramebuffer no se serializa en savestates (verificado: ningun
  DoState ni copia cruda lo toca; los vfbs se reconstruyen al cargar).

+++ GPU/Common/Draw2D.cpp
@@ -309,7 +309,8 @@ Draw2DPipeline *Draw2D::Create2DPipeline(std::function<Draw2DPipelineInfo (Shade
 	};
 }
 
-void Draw2D::Blit(Draw2DPipeline *pipeline, float srcX1, float srcY1, float srcX2, float srcY2, float dstX1, float dstY1, float dstX2, float dstY2, float srcWidth, float srcHeight, float dstWidth, float dstHeight, bool linear, int scaleFactor) {
+// STV_ESCALA_v1: scaleFactor float de punta a punta (ver Draw2D.h).
+void Draw2D::Blit(Draw2DPipeline *pipeline, float srcX1, float srcY1, float srcX2, float srcY2, float dstX1, float dstY1, float dstX2, float dstY2, float srcWidth, float srcHeight, float dstWidth, float dstHeight, bool linear, float scaleFactor) {
 	float dX = 1.0f / (float)dstWidth;
 	float dY = 1.0f / (float)dstHeight;
 	float sX = 1.0f / (float)srcWidth;
@@ -330,7 +331,7 @@ void Draw2D::Blit(Draw2DPipeline *pipeline, float srcX1, float srcY1, float srcX
 	DrawStrip2D(nullptr, vtx, 4, linear, pipeline, srcWidth, srcHeight, scaleFactor);
 }
 
-void Draw2D::DrawStrip2D(Draw::Texture *tex, const Draw2DVertex *verts, int vertexCount, bool linearFilter, Draw2DPipeline *pipeline, float texW, float texH, int scaleFactor) {
+void Draw2D::DrawStrip2D(Draw::Texture *tex, const Draw2DVertex *verts, int vertexCount, bool linearFilter, Draw2DPipeline *pipeline, float texW, float texH, float scaleFactor) {  // STV_ESCALA_v1: scaleFactor float
 	using namespace Draw;
 
 	_dbg_assert_(pipeline);
@@ -345,7 +346,7 @@ void Draw2D::DrawStrip2D(Draw::Texture *tex, const Draw2DVertex *verts, int vert
 	Draw2DUB ub;
 	ub.texSizeX = tex ? tex->Width() : texW;
 	ub.texSizeY = tex ? tex->Height() : texH;
-	ub.scaleFactor = (float)scaleFactor;
+	ub.scaleFactor = scaleFactor;  // STV_ESCALA_v1: ya es float, sin cast que trunque
 
 	DepthScaleFactors zScaleFactors = GetDepthScaleFactors(gstate_c.UseFlags());
 	ub.zScale = zScaleFactors.Scale();
